modularised the schema and completed JsonScheme.php. might want to remove the last line
